refactor: use only os.GetEnv and move it into utils package
